chore: SVGR 적용전
