--- src/components/ui/Dropdown/index.ts
@@ -1,3 +1,0 @@
-export { default as ModalDropdown } from './ModalDropdown';
-export { default as FilterDropdown } from './FilterDropdown';
-export { default as SortDropdown } from './SortDropdown';
