--- src/components/ui/Dropdown/types.ts
@@ -1,6 +1,0 @@
-export type DropdownProps = {
-    options: string[];
-    selected?: string;
-    onSelect: (value: string) => void;
-    placeholder?: string;
-};
